Improve OpenCL logging and warnings
Explicitly catch and report the 'Unhandled' and 'OpenCLError'
exceptions, too, in DynamicKernel::create, instead of relying on a
catch(...).

(Actually not sure why we need three separate classes of OpenCL
handling exceptions.)

We have the code to show the file and line where something exceptional
(or actually not so exceptional) happens with OpenCL, so make use of
that then.

It is a perfectly normal thing to throw (and catch) at least the
UnhandledToken and Unhandled exceptions (when compiling a formula and
noticing that we don't want to use OpenCL for it after all), so it
does not make sense SAL_WARN about them. Just SAL_INFO is fine.

Change-Id: I56770f47622cb4bc40f2216c5e78d3da3c7c9adf