tdf#90697 docx - don't change continuous break into page break
As soon as you set PROP_PAGE_DESC_NAME, you are inserting that
style as a page break.  Setting a pagebreak via a continous break
was first introduced in commit 79d39c064ef2da77226749d524519b9a851e2653
but the unittest for that commit no longer uses this code.
I'm suggesting it be reverted.  It really messes up round-tripping
when continuous breaks are removed/replaced with hard page breaks.

There are a few odd cases where the very first section needs to set the
page break via the continuous break, so it hasn't been eliminated
completely.

Change-Id: I5b0e4bb7612ee9df47e5c49b8c2316dc001b5238
Reviewed-on: https://gerrit.libreoffice.org/26662
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>