uitest: add a way to have sleeps in some test runs
Now adding a -d or --debug to the command line options will enable the
sleeps in the uitest.debug part. This makes it much easier to debug a
test in the visual mode.

Only remaining part is now to add a test runner into the build system
that makes use of that mode.

Change-Id: I03d55b10f06dd12a63a8d87c135967901bef0fba