ODT export: add embedded pdf support
Once a .pdf file is inserted in Writer using Insert -> Image, Writer now
advertises a replacement metafile for it (to make LO <= 5.2 and other
ODF readers still be able to read the graphic, where the replacement is
the same as the original metafile, but without .pdf data), and the ODF
export writes the .pdf data with the correct extension / mime type.

Also extend the checksum logic to take care of the pdf data, so on
export xmloff won't think that the full and the replacement graphic are
the same.

The import side is still missing, so no testcase yet.

Change-Id: I5c5b7c80ca4024ecbcb5b2d4442d21ca33755546
Reviewed-on: https://gerrit.libreoffice.org/26695
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>