Bad cast from SanExtensionImpl to CertificateExtension_XmlSecImpl
...both ultimately derive from css::security::XCertificateExtension, but that is
all they have in common.  The special handling of 2.5.29.17 (and thus the bad
casts) was introduced with 35580249574120ec222d01f941ba314493076c05
"tkr38: #i112307# Support for x509 v3 Subject Alternative Name extension added".
Lets assume that it was an oversight there that setCertExtn (which is a function
of CertificateExtension_XmlSecImpl, not inherited from a---common---base class)
should not be called on such special-case SanExtensionImpl instances.

Change-Id: I96cfd42f606c79920d1548f323b68f17ff4e9081