tdf#96015 (part) links to Askbot in the Help menu
Add the entry "~Get help online" to the modules
Help menu.

on click, the system browser opens
http://hub.libreoffice.org/forum/?&LOlang=<lang>

<lang> is obtained from .getLocale()

That in turn will redirect to e.g. askbot (but only few languages are
using it at the moment) or other nl-specific resources.
(nl-projects can request corresponding targets from infa, without the
need to change the code – cf rdm#1983)

The link can/should be extended to also include the version of LO and
the currently used module to allow finer grained redirects.

Change-Id: I56108dba4dd2684405623c4586103fc8ff9bea56
Reviewed-on: https://gerrit.libreoffice.org/27028
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>
Tested-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>