Related: tdf#100861 same selection recorded multiple times...
in FindAll libreofficekit impress test

on find all we loop through the textboxes searching for the string.

We start by searching into the first textbox with the string in it.
mbStringFound gets set to true and this first selection is reported.
Now the current pos is still in that textbox at the end of the string.

The next loop will find nothing in this textbox, but because mbStringFound
was set in the earlier pass, the same selection gets reported again.

The next loop will move to the next textbox.

To keep this fix as simple as possible just check if the selection was
the previously reported one and skip it if it is.

I believe this is the problem that

commit 8cb5e7c005759c3da9ce39fdf804a2502d2f6510
Author: Marco Cecchetti <marco.cecchetti@collabora.com>
Date:   Mon Jan 11 16:43:02 2016 +0100

    lool - search all - unit test failure - solved

wanted to solve

Change-Id: I30e7b9c581488b48fa27f138209f291063b459a3