Revert "bccu#1781: Impress/Comments - selection not updated on delete"
because selection is visually mangled in the editeng.

type a word in calc, hit return, double click on the word, then end, hold down
the shift key and move to the left, every second letter gets highlighted.

The original bug is private so I don't know what problem it wanted to fix
to see if there is a better solution.

This reverts commit 5caf35679b1689f3920a93fe4da03e6a21ec78aa.

Change-Id: Ib1b535aca98daaae210db4e3d62900fbc389e816