screenshots: Use UI string and fallback for UI-String only
Two changes in this commit:
The Dialogs to be dumped are identified in their test files
using their UXMLDescription to allow later to 'find' the
known dialogs and use the specialized construction for these.
Also added a fallback to construct a vcl Dialog based on only
the UXMLDescription and the VclBuilder. This will be constructed
without any active initialization/layouting, so should only be
used for unknown Dialogs. Also added a dumpDialogToPath version
to the tooling that can work directly with a vcl Dialog instead
of a VclAbstractDialog.

Change-Id: I90abb6f59c2fcc5d534907ae7e4b9a15edc2d694