xmloff: forms import: convert relative xlink:href to absolute
There is currently only one place in the forms directory that converts
relative hyperlinks to absolute on import, in
OURLReferenceImport::handleAttribute(), but there are other elements
that have xlink:href attribute as well, such as form:form.

The export of form:form xlink:href does convert absolute to relative,
in exportTargetLocationAttribute(), but if the model URL is actually
already relative it will be converted to absolute instead, oddly enough.
This leads to different absolute href attributes, depending on the
directory where the file is exported, as can be observed with e.g.
ooo95698-1.odt and fdo40634-2.odt.

Let's apply a big hammer and make all "href" attributes absolute on
import.

Change-Id: I39d05707f3a8a899c7bbde8d9c0e2bc006c39e12