No need for own implementation of erfc() in OpenCL
The own code was copied from the C++ one we used to have in
sal/rtl/math.cxx but which was removed in
3f631902a77eb2490b6d0fa868cab94fc96622f1. However, it did not work
correctly on some machines at least, like my AMD A10-7800 running
Windows 10. I was unable to figure out why not. This lead to OpenCL
being disabled by the Desktop::CheckOpenCLCompute() code we now run
early on startup.

Anyway, as OpenCL has erfc(), just use that.

Change-Id: I7ba6104fc4975cd570358760fa97a19390a54cce