tdf#91665 l10n: Replace string list by separate strings in ScPrintUIOptions
String lists are sometimes (mis)used to group unrelated items;
the advantage(?) being that only one identifier needs to be defined
for the whole group. The items are then referenced by index,
which is rather fragile when they are used in different parts of UI.
String lists like this one should be replaced by separate strings.

Change-Id: I78a2c8d1a8e5b1fed6a2d0afefd96ec126bb79af
Reviewed-on: https://gerrit.libreoffice.org/27689
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>