Reduce ENABLE_OPENGL / ENABLE_HEADLESS confusion
ENABLE_OPENGL means whether to enable the OpenGL slideshow transition
code. It does not mean whether to enable use of OpenGL in general. So
rename it to ENABLE_OPENGL_TRANSITIONS while at it.

ENABLE_HEADLESS means whether to disable use of X11 and OpenGL on X11
(and Wayland) platforms, I think, meaning Linux and maybe Solaris and
the BSDs. Maybe it should be renamed to DISABLE_X11_AND_OPENGL.

Change-Id: Ibb30f51646b1bcc477fe691a3fa38c7a1e3944ae