Resolves: tdf#92530 put the rest of missing validity tips into the other grids
this stems from...

commit fe13a763a61081044eb414051daa599311c1fa90
Author: Kohei Yoshida <kohei.yoshida@suse.com>
Date:   Thu Jan 5 22:13:52 2012 -0500

    fdo#36851, bnc#737190: Make the data validation popup more reliable.

    Previously, the data validation message popup was not sticking on
    reliably on Linux, and flickers very bad on Windows.  That was due to
    the painting conflict between the top-most frame window and the grid
    window.  By making the message popup window a child window of the
    grid window, it paints more reliably.

where the tip was resticted to the current grid because of problems from

svx/source/sdr/overlay/overlaymanagerbuffered.cxx and
OverlayManagerBuffered where the original flickering window isn't handled
nicely by the background save/restore stuff there.

So here its all been reimplemented as an overlay which plays better
with that code, and if it doesn't fit, then its in all grid windows,
so visually it appears as a single note that floats underneath the
split window lines.

Change-Id: I18ac42c2a0686ab509252830741d36fb17cae629